Shield Update
Created a shield object that the player can use to protect himself against projectiles. It breaks after one use to tie into the one-time use theme.

Created a "RotateOnGrabLeftHand" Script that checks if the object was grabbed with the left hand and sets the grabpoint to a new rotation. So if objects need to be held differently on the left and right hand. Create them like their meant to be for the right hand and set the proper rotation on that script for the left hand.

On the Player the Left Base Controller got a tag called "LeftHand" which makes the rotation script distinguish between them.