#28 Item Spawner
I have created a few SpawnZones for the items. You can find the items under Resources/Items (still dummy assets, later please name the models the same way).

It is now possible to select the desired item in an item spawn zone (script), how many of them should be created and in what radius they are spaced apart.

The Item Spawn Zone will then communicate with the ItemManager, which will instantiate the items at the desired spawn zone.

When the items are created in the spawn zone, the DestroyOnStandStill script is added to an item. If the item is inside the Box Collider of the SpawnZone, the DestroyOnStandStill script is disabled, on leaving it is enabled.

When the boss loses a body part, the event onDestroyLimb can be triggered. Here then every spawnzone is deactivated, which has selected the passed item type.