Fixed Projectile Spitter / Edited Laser On player found
Projectile Spitter Projectile Pooling threw null exception when the projectile was destroyed. Commented out the pooling, now back to instantiating and destroying projectiles which works.

Laser threw null exception when the laser started to prepare an attack and the boss lost vision of the player.
Commented out the on player lost target = null funciton.